+++ yax/postgres/src/main/scala/doobie/postgres/hi/largeobject.scala
@@ -1,8 +1,8 @@
-package doobie.contrib.postgresql.hi
+package doobie.postgres.hi
 
 import doobie.imports._
 import doobie.util.io.IOActions
-import doobie.contrib.postgresql.imports._
+import doobie.postgres.imports._
 
 import java.io.File
